renamed shit because I realised the old names where kind of dumb
