--- src/main/kotlin/dev/avt/api/person/absentie/registerHours/RegisterHoursRequest.kt
@@ -1,6 +1,0 @@
-package dev.avt.api.person.absentie.registerHours
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RegisterHoursRequest(val token: String, val hours: List<Int>)
