+++ src/main/kotlin/dev/avt/api/person/absentie/requestHours/RequestHoursRequest.kt
@@ -1,0 +1,6 @@
+package dev.avt.api.person.absentie.requestHours
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestHoursRequest(val token: String, val hours: List<Int>)
