+++ src/main/kotlin/dev/avt/database/RequestedHoursSchema.kt
@@ -1,6 +1,5 @@
 package dev.avt.database
 
-import dev.avt.database.MagisterData.Companion.referrersOn
 import kotlinx.coroutines.Dispatchers
 import org.jetbrains.exposed.dao.IntEntity
 import org.jetbrains.exposed.dao.IntEntityClass
@@ -11,22 +10,22 @@ import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
 import org.jetbrains.exposed.sql.transactions.transaction
 
-class RegisteredHoursTable(id: EntityID<Int>) : IntEntity(id) {
-    companion object : IntEntityClass<RegisteredHoursTable>(RegisteredHoursService.RegisteredHours)
-    var user by AVTUser referencedOn RegisteredHoursService.RegisteredHours.user
-    var hour by AvailableHoursTable referencedOn RegisteredHoursService.RegisteredHours.hour
+class RequestedHoursTable(id: EntityID<Int>) : IntEntity(id) {
+    companion object : IntEntityClass<RequestedHoursTable>(RequestedHoursService.RequestedHours)
+    var user by AVTUser referencedOn RequestedHoursService.RequestedHours.user
+    var hour by AvailableHoursTable referencedOn RequestedHoursService.RequestedHours.hour
 }
 
-class RegisteredHoursService(database: Database) {
+class RequestedHoursService(database: Database) {
 
-    object RegisteredHours : IntIdTable() {
+    object RequestedHours : IntIdTable() {
         val user = reference("user", UserService.Users)
         val hour = reference("hour", AvailableHoursService.AvailableHours)
     }
 
     init {
         transaction(database) {
-            SchemaUtils.createMissingTablesAndColumns(RegisteredHours)
+            SchemaUtils.createMissingTablesAndColumns(RequestedHours)
         }
     }
 
@@ -34,6 +33,6 @@ class RegisteredHoursService(database: Database) {
         newSuspendedTransaction(Dispatchers.IO) { block() }
 
     companion object {
-        lateinit var INSTANCE: RegisteredHoursService
+        lateinit var INSTANCE: RequestedHoursService
     }
 }
